Redesign term representation to encode type at compile time.
The representation of terms is now a type 'Term (f :: * -> *) :: *',
where 'f' is the representation of the contained type; that is, if
a term of type 'Term f' represents a value of type 'a', then
'f = Rep a'. This has two benefits:

* Terms (and the compiler) are aware of their type. We don't need a
separate wrapper structure which stores a runtime representation of
the type.

* All of the metadata of 'Rep a' is encoded in the type of the term,
so we can reconstruct the original Haskell value from a 'Term'
representing that value.

The downsides of this are twofold:

* It is horrendously non-portable (then again, so was the prior
implementation)

* It is horrendously complicated.

Both of these problems will be mitigated by an abstraction layer
which allows plugin implementations, a portable one and a non-portable
one with extra features. Coming soon.