Redesign the representation of terms to avoid non-portable generics.
This is a big pivot. Instead of using GHC.Generics to represent
deconstructed Haskell types, we use a more ad hoc approach. Primitives
are stored as constants, lists are stored as lists of terms, tuples
are represented using a "Product" term constructor.

ADTs can be constructed by storing an ADT data constructor along with
a term which, when unified with a constant, can be passed to the
constructor to create the ADT value.

This system is much better than the old direction: it's far more
portable and way easier to understand.