Implement the main proof engine.
The solver is capable of finding all proofs of a given goal which
follow from a given set of clauses. The interface of the solver
consists of a single entry point returning an opaque structure,
together with a handful of functions for extracting information
from that data structure. This allows the client to choose the type
of information and level of detail they want.