Implement an interactive debugger for HSPL programs.
This is accomplished by altering the main solver algorithm to use
a continuation-based control flow which allows third parties to
hook into interesting events during the runtime.

The debugger is one such party. It currently provides basic
navigation commands, such as step, next, and finish. The mechanism
is general enough to allow the implementation of many more commands,
including printing the goal stack, printing the list of alternatives
of the current goal, forcing the current goal to succeed or fail, etc.
However, the basic functionality provided here should be enough to
aid in the development of nontrivial HSPL programs, which can in turn
be used to stress the prover.