Implement the equivalent of Prolog's =/2.
The goal (|=|) acts as a predicate with special semantics: it
attempts to unify its operands, and succeeds if and only if the
unification succeeds.

Implementing this required a significant overall of the AST and
solver. We introduce a new notion of a Goal, which is any expression
that can appear on the right-hand side of (|-). A goal can be a
predicate, but it can also be a unification goal. This extra level
of abstraction will make it easier to add more special goals in the
future.