Implement the equivalent of Prolog's ==/2.
This predicate (|==|) succeeds if and only if both of its arguments
are identical; that is, they have already been unified.

In order for this to work, the intermediate unifier resulting from
proving a subgoal must be applied to the next subgoal to be proven,
and so on. It was actually a bug not to do so, but that is fixed now.