Fix an issue where clauses are added to programs in the wrong order.
Ast.addClause was consing new clauses onto the front of the clause
list for each predicate. This leads to confusing behavior when the
order in which alternatives are tried matters. For example, recursive
predicates must declare the base case _after_ the recursive call; to
do otherwise leads to left-recursion.

To make this more intuitive, we append new clauses to the end of the
list, and so alterntives are tried in the order they are defined in
the source file.