Consolidate superclasses of HSPLTypes into a single type class.
Any type which is to be represented by a term is required to have
an instance for Data, Typeable, and Eq, as well as Show if the
ShowTerms flag is enabled. These requirements were repeated (often
inconsistently) across the codebase. Replace all the complicated
constraints with a single 'TermEntry a' constraint, where TermEntry
is a subclass of Data, Typeable, Eq, and Show, and it has an instance
for any type with an instance for those superclasses.