Fix an issue where certain unifications were lost by the solver.
The affected case is a predicate goal with one or more subgoals. In
this case, each subgoal is proved recursively, and the resulting
unifier is propagated through and applied to each goal before proving
it. However, the initial value for the unfier was mempty, rather than
the unifier created from unifying the goal with a horn clause. This
caused any substitutions in that unifier to be lost. The fix is to
simply change the initial value.