Change the representation of ADT terms to facilitate generic manipulations.
Specifically, we replace tuples with type-erased lists.

Tuples are a PITA, so let's get rid of them as early in the pipeline as
possible. Of course, getting rid of them turned out to be an even bigger
PITA, but now that's done and it's much easier to write generic functions
over type-erased lists than tuples.

Of course, we sacrificed some degree of type safety through the erasure,
and there are more calls to 'error' in the codebase than there were before.
In theory, though, those errors should never surface, because the higher
level API which does the type erasure is itself typesafe.