Add a default 'toTerm' instance for ADTs.
Uses GHC.Generics to construct a generic implementation of 'toTerm'
for algebraic data types. Because all types that we may want to put
in a term now have a 'toTerm', 'Termable' (the class of 'toTerm') is
a super class of 'TermEntry'. This required some shuffling around of
the Term-related classes. We can't have TermData => TermEntry and vice
versa, so we have a new class Termable containing the method toTerm,
and a class TermData which is a subclass of 'Termable a' and
'TermEntry (HSPLType a)'. For the purposes of writing constraints in
type annotations, 'TermData' is roughly equivalent to the old version.
Now, users can provide instances of Termable (not TermData) to use
ADTs with HSPL.