Improve semantics of proveEqualWith
Since all 'TermEntry's are now 'Termable', we can express the
semantics the 'Equal' goal in a much simpler way. Modulo Maybes, the
evaluation of the rhs can be approximately expressed by 'toTerm . fromTerm'.
This is both simpler and more general than explicitly evaluating arithmetic
expressions. For example, we can now evaluate subexpressiosn (in ADTs eg)
and unify the resulting simplified ADT with an ADT containing a variable on
the lhs.