Implement a more complete logical language in Goal
This patch adds And, Or, Top, and Bottom goals in addition to the
existing Not connective. This allows us to express many first-order
logic sentences as complex goals. In particular, the conjunction which
forms the right-hand side of a predicate definition is now expressed
using And, rather than a list of goals.

This new expressivity allows users to nest disjunctions in the right-hand
side of a predicate definition (similar to Prolog's ';'), without having
to expand the definitions into multiple Horn clauses.

We also provide syntactic sugar for Top (true) and Bottom (false).