Add a goal analagous to Prolog's findall/3.
An Alternatives goal takes a template term, a goal, and an output
term. At runtime, it executes all branches of the goal to obtain a
list of proofs and unifiers. It then applies each unifier to the
template goal to create a list of instantiated templates. It attempts
to unify that lit with the output term, succeeding or failing accordingly.