Create a module dedicated to displaying data structures.
The purpose of the module is to isolate code which turns internal
data structures into user-facing strings. Giving this code its own
module also makes it a bit more convenient to apply some more complex
logic to generate nicer results, compared to doing it all in the various
show instances. We can now also use the show instances as they are meant
to be used, to print something close to the Haskell code representing an
object, which is useful in development and debugging.