Factor out common implementation of Unifier and Renamer.
Both of these types map (forall a. TermEntry a => Var a) to
(forall a. TermEntry a => f a), where f is either Term or Var. Worse,
they have a complicated implementation without a real interface, making
for very clunky code.

This patch creates a module dedicated to this data structure (no
parameterized by f) and exports an interface similar to that of
Data.Map, greatly simplifying the unification code.