Simplify logic state interface.
Rather than having two state types and objects, one for backtracking
state and one for global state, we only have one state. The state must
be an instance of a class that lets us split the state into global and
backtracking components, and then later recombine those components.
With that, since we have only one state object, we can implement the
regular MonadState class and do away with MonadLogicState.