Make goal-tracking optional.
Currently, we record in memory every single goal that is proven
during the course of a proof. This takes a tremendous amount of
memory to store information 99% of which will never be queried.
Thus, we introduce a new goal, Track, which modifies a goal indicating
that it's result should be recorded and made available via queryTheorem.
Any goal not modified by Track is not recorded.

There are two syntaxes for tracking a goal. The first is a function track,
which takes and returns a GoalWriter. This is used to track goals at the
invocation (most granular level). A less granular tracking can be achieved
with the predicate attribute Theorem, which automatically tracks every
invocation of a predicate.

Predicate attributes are also a new concept: they are simply flags which
modify the behavior of a predicate. In addition to Theorem, we have
SemiDet, which replaces the old semiDetPredicate constructor. This is a
more scalable way of modifying predicates than having, say, predicate,
semiDetPredicate, theorem, semiDetTheorem, etc.