Optimize Unifier composition.
Composing Unifiers everytime we complete a new subproof is expensive,
in terms of both time and memory. To mitigate this cost, we implement
to optimizations:

1. Composition is lazy. Unifiers are never explicitly composed, only
unioned. Whenever a unifier is _applied_, we apply it recursively to
the result until there are no more unifications to be made. This
achieves the same transitivity that composing the unifiers does.

2. VarMap union is made more efficient by implementing it in terms of
Data.Map.union.

These optimizations yielded a speed up of 30-100x.