Factor out operator preferences.
All operator preferences are defined in the same part of the file,
making it easy to see the relative order. Also, we use a macro
PREC which is continuously redefined to decreasing integer constants,
which will make it easy to move operators to a different precedence
level in the future with minimal code change.