Tag predicates definitions with their location in the source code.
This allows us to relax the constraint that predicate names be
unique quite alot.

There is a caveat: source location is not available prior to base
4.8.1, so with older versions of base names will still have to be
unique. To alleviate the user's burden even in this edge case, we
provde a mechanism for tagging predicates with an optional scope
string, and require uniquencess only among predicates in the same
scope.

Fixes #17