Expose primitive goals with a wider span.
We add two new goal constructors: Once (again) and If. These make
it possible to implement a variety of high-level sugar, including
once, ifel, not, .||. (short-circuiting or), and cond without
relying on cut, thus removing the confusing cut semantics of once
and cond. We can also remove the now-obsolete Not goal constructor.

Closes #24